pyproject.toml
[project]
name = "documentrag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.14",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "python-dotenv>=1.2.2",
]
